Force Desaturation Algorithm
cGenericTool.cpp
- created a desaturation method.
-this method checks the goal to proxy vector and evaluates if it is in
the saturation zone. If it is, it moves the global position of the goal
in the direction of the proxy to bring it back to operating limits of
the device.
-desaturate is called in the updatePoseNoisy after the global position
vector is calculated. after the cursor position is desaturated, then we
add the noise to the cursor position.

cAlgorithmFingerProxy.cpp
- created computeProxy method
- this method is just the first half of computeForces
- it uses the goal position to find the proxy position and returns the
new proxy position.