Updated the desaturation algorithms to also desaturate the no noise case. Increased Cursor Size. Reduced noise generation frequency to 50hz
Desaturating the no noise case allows us to better feel ourselves
dropping into the hole.

11/27/17
-tuned parameters for cursor size
-due to increased frame rate drop for a large cursor, reduced the noise
generation rate to 50hz

GenericTool.cpp
- updated the updatePoseNoisy algorithm to switch desaturation methods
based on whether there is noise in the system using a boolean passed to
it which is passed to desaturate()
- updated desaturate to also take them same boolean and change the
desaturation.

Phantom.cpp
- updatePhantom()
-now uses updatePoseNoisy() for all pose updates. if the noise toggle is
off, we do not do any noise calculations and just use noise = 0.